Use climada develop branch while bug exists on main
Since Stamen stopped providing free, unauthenticated access to their
map tiles, any code that explicitly refers to it has stopped working.

This includes the CLIMADA stable release Impact class, which is
imported when loading the storm_europe impact function.

The problem has been fixed on the CLIMADA develop branch. So we'll
use that for development for now.