Optionally toggle BI in direct.py. Never apply to agriculture
The new APPLY_BUSINESS_INTERRUPTION variable is declared at the
start of direct.py which lets you turn on and off the conversion
of % asset loss to % production loss. Mostly for debugging, at
least for now.

Also, remove the impact function 'fallback' we provided when
nothing was defined: this was the default CLIMADA TC. Now we
throw an error when no impact function is defined: we should
always have one.

Move the storm_europe impact function from storm_europe into
direct.py. Seemed a smarter place to keep it. Maybe we'll put it
back sometime. Who knows.