Bugfix agriculture region_ids
I was running into an issue where the CLIMADA API's downloaded
hazard had all centroids with region_id set to 1. This meant that
we didn't calculate our own and then all centroids were removed
when we filtered to a region.

Given that this didn't happen in previous runs of the model, I
assume it's to do with the recent change in the Centroids class
to be a geodataframe.

The quick and dirty workaround here is to check if the centroids
have a geodataframe, and if it's got region_ids all set to 1. If
so, to set them to nan and continue as planned.