Add economic_assets as a sector
This sector uses vanilla LitPop as its exposure and doesn't
apply any business interruption.

The results from the direct impact calculations will be useful for
validation and calibration. The results from indirect impact
calculations will fail, since this isn't an MRIO sector.

i.e. this sector is designed to be called with functions
within the pipeline.direct package and not part of a full supply
chain run.