Big analysis pipeline refactor
Big changes to the calculation pipeline.

1. Calculations are orchestrated through a dataframe of calculation metadata

Main conceptual change is that all required calculation metadata are stored
in the analysis_df dataframe, but no data are. This allows us to set up
all the calculations in advance and then loop through them (with the option
of paralellising). All outputs are written to file, meaning that large
calculations don't need to be stored in memory, and so that calculations
can be started and stopped. If a file has already been calculated it is
loaded from memory instead of being rerun (unless force_recalculation is set
to True).

Note: this dataframe is only created for the individual
scenario-hazard-sector-country combinations, because all of these need to
be calculated before any aggregated calculations can be set off.

2. Aggregate impacts into a new COMBINED hazard

This commit also introduces an extra stage to the calculation where
additional calculations for the _sum_ of the impacts across hazards (i.e.
'multihazard') are calculated for each scenario-sector-country. This
inovles some extensions to CLIMADA's yearset functionality to combine
impact objects that are also yearsets.

3. Make windstorm work properly

This should have been added to the codebase a while ago but for some
reason it's only coming now: wind storm hazards need to be aggregated from
daily hazard to yearly totals before they are fed into risk calculations.

4. Metadata

We also write some information on what calculations were performed and
which succeeded to csv files called direct/calculation_report.csv and
indirect/calculation_report.csv in the relevant results directory.