Add scripts to generate intermediate SLR impact data
Without a probabilistic sea level rise dataset we're producing an intermediate
product that overlays the WRI Aqueduct coastal flooding risk return period maps
for different scenarios and return periods with the different exposures used in
our analyses.

The methodology assigns each grid cell in the SLR risk map to the exposure grid
cell that it's within (since the Exposures are lower resolution than the SLR
map). Then it takes the maximum flood depth for each exposure as its result.

These exposure-flood depth mappings are written out, and summary statistics of
affected percentages by country are created.

The method is slow and computationally expensive. There's definitely a smarter
way to do this by converting the CLIMADA exposures to rasters, but I leave that
for a smarter person to do.

This add scripts to resources/impacts/slr_intermediate:
- 01_download_aqueduct_data.py:
Downloads the raw data from the WRI site. Exposure data is assumed to exist and
is accessed through the CLIMADA Data API or (currently) the team's S3 bucket.
Writes metadata for the files to data/aqueduct_download_metadata.csv
- 02_calculate_bounding_boxes.py
As preparation, calculates a rough bounding box for each country that we'll be
working with and stores it in a CSV. (Could actually be removed since the next
script also does something similar, but the chain was working so I haven't done
this). Writes to data/country_bounding_boxes.csv
- 03_calculate_exposure_metadata.py
As preparation, calculates summary stats for each sector-exposure combination,
including total economic value and the inferred raster metadata for its
underlying grid. Writes to data/exposure_metadata.csv
- 04_calculate_slr_intermediate.py
The main calculation. Loops through each sector-exposure combination of exposures
and through each sea level rise scenario-year-return period hazard map to calculate
the maximum flood depth at each exposure point and writes results to
data/results/*
- 05_summarise_outputs.py
Summarises the above stats to give, for each combination of country, sector,
scenario, return period, future year and flood depth, the total value of assets
affected, expressed as a dollar value and a % of total assets