Update requirements, README, tests that it all runs
See the new README.md for an explanayion of the install instructions.

Move requirements information into a requirements folder, now with
three files:
- requirements.yml: for an out of the box install
- requirements_jupyter.yml: to add Jupyter notebooks on top
- requirements_dev.yml: for advanced users. See README

Also adds a unit test for analyis.py that runs a very simple pipeline
so that you can check the code runs with a simple
   python -m unittest

And the configuration file needed to run it.