Improve logging of which supply chain calculations we need to request in analysis_df
This gives a bit better acccounting fior which calculations and outputs are and aren't done
and is analagous to the other phases of the calculation. Separate columns dor leontief and
ghosh outputs are created.

Also means we don't recalculate existing data if a previous version of the calculation was
stopped part way through.