Add timestamp to output summary filename
The analysis chain writes a csv giving info on what it wanted to
produce and what it was able to produce. Since the analysis.py
might be run multiple times we decided to put a timestamp in the
filename.

In addition this commit only writes the file out to one location,
the indirect_output_dir. Earlier a copy was written to the direct
and indirect dirs, but that's a bit confusing.