If file already exists, don't download from S3
During the calibration we run the same code many times and don't want to
redownload things every time. If the target file already exists on the
local file system, it's not downloaded from S3.

Open for discussion about whether this check should be inside the
download method or somewhere else!

Also added the 'overwrite' parameter to the download method that lets you
download even if the file exists locally. It's not used anywhere but we
could potentially add it to the config.

Also a little bugfix where downloading to a target path in the current
directory threw an error when creating the target download directory.