Change TC data source for new improved data
Thanks to Samuel Juhel for generating awesome new TC data for us.

The new files are stored on the S3 bucket and this commit points all
tc hazard requests to the bucket.

Note: you may want to clear your local CLIMADA cache which is probably
storing all of the old TC files that we won't be using any more.

The cache is by default located at
~/climada/data/hazard/tropical_cyclone

The files here can safely be deleted.