Small refactor to direct impact function construction
For the calibration we want to inject custom impact functions.

This allows that – the config parameter 'calibrated' can now take
other values, which prompts the script read a custom.csv file
from the relevant folder in the resources directory instead of
using a hard-coded function.

The implementation here is partial to get the calibration of
direct hazards working. Something analagous will be added later
for BI.

Note that the construction of the impact function has changed
slightly: analysis.py now gets the asset loss impact function
and identifies the BI function and composes them, instead of
calling a hazard-specific function that does this.