Package restructure for pip installation
It was getting too complicated to get the pip setup and install to work so
this commit restructures the code to look like most python packages.

- Everything that we install has moved into a new ./nccs/ directory - i.e. the
pipeline, analysis, dashboard, utils, etc. Internal code that isn't part of the
'published' nccs utilities (data, resources, code to generate data, etc) stays
in the root directory. Some of these decisions were pretty subjective and I'm
sure we'll want to change our minds on what is and isn't included in the install.
- All the import paths I could find are updated, e.g. import pipeline.direct becomes
import nccs.pipeline.direct
- The util folder_naming has had its paths updated.

A restructure like this will cause a whole bunch of future errors. I apologise.