Refactor calibration module for speed and being generally better
- Calibration creates lots of model run outputs. If a previous calibration was interrupted
we can read in the old output and carry on where we left off
- Often our model pipeline results are linear in a parameter - usually our impact function
has some vertical scaling. Given one model output for this we can create hundreds of
variations on it basically for free. Now when specifying your calibration input you can
specify a linear_param, and the calibration takes these free wins as a way to speed things
up.
- Calibration now has a cache where existing results are stored. If an existing result is
requested, or an existing result differing only by the linear parameter, we can simply
read it from the cache.
- Loads of debug logging (sorry)
- General quality of life: info on timings etc. Logging in general