Update TC observations: missing countries, RPs, RP loss by region
This finished the work started in the previous couple of commits
- some countries were missing from our TC hazard event set, we have them
now and can update the observation column that indicates whether the
events are modelled in CLIMADA (not currently used for calibration)
- Some RPs were incorrectly assigned. This corrects them
- Instead of producing the output tc.csv with rows assigning a return
period to each country-event entry, we output tc_country.csv with these
country-event RPs and tc_region.csv which instead assigns RPs to each
region-event. We think this might lead to a better calibration.