Bugfix yearsets
The implementation of yearsets had an error where event sets where each
event was one year of hazard (e.g. crop yield) were being sampled with
a Poisson process, when actually they should have always been sampled
with one event per yearset year. This is fixed.

Rather than trying to infer whether the yearsets should be modelling a
Poisson process (it's usually the case if the event frequencies all add
up to ~1), I've added it as a parameter to the method. This seems
more secure.

At the same time a few other changes to yearsets:
- Yearset creation and combination allowed for individual impacts to
be capped at the exposure level, either to a scalar value or according
to an Exposures object (i..e. not more than 100% loss). This wasn't
properly implemented, and only changed the Impact object's impact
matrix. This updates the calculation so that all the properties are
updated. From a quick look through the results, it shouldn't really
change anything.
- Change the default yearset impact cap from 1 to None. Seemed a bit
dangerous to cap at 1 if we don't know the impacts have already been
normalised. Doesn't affect any code because the value is always
specified.
- Update tests