Bugfix: combining impacts when some values are missing
The order of elements in a dictionary matter when it's converted to a
pandas Series and then a DataFrame. This fixes a bug that occurred when
combining impacts if one or more of the impacts was missing. In this case
a dictionary object was created with elements in a different order.

It should probably be named list or structured dict or something but this
works.